requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.3.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.2.0
requests==2.31.0
soupsieve==2.5
sympy==1.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
